+++ eslint/jest.config.ts
@@ -1,4 +1,5 @@
 export default {
+  resolver: '<rootDir>/tests/jest-resolver.ts',
   preset: "ts-jest",
   testEnvironment: "node",
   testMatch: ["<rootDir>/**/tests/**/*.spec.ts"],