+++ eslint/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# 0.3.0
+_9 February 2022_
+
+* Removed unused peer dependency `eslint-config-prettier`
+* Disabled a few rules that were more annoying than useful:
+  * `capitalized-comments`: Annoying when commenting out code temporarily, ends up adding a lot of work to fix the code when it is uncommented again.
+  * `@typescript-eslint/naming-convention`: There are many edge cases where this rule has to be disabled. It is also difficult to configure since it requires copying the entire rule configuration to make modifications.  
+  * `@typescript-eslint/no-non-null-assertion`: Null assertions should be avoided, but they are useful sometimes when it is clear from the code that the value cannot be null/undefined, but the interpreter is unable to infer it.
 
 # 0.2.0
 _28 October 2021_